Add startTls() method
Applications can invoke this method to decide if STARTTLS command should be issued.

The admin can override the default: if a sssd/StartTls prop is defined with value "enabled" or "disabled" the method returns true/false accordingly.

By default it returns true if a remote LDAP provider is configured, bind credentials are available and SSL is not enabled.